Remove package objects in cmtc and cmta
- Package objects will be deprecated and can be replaced by
  Scala 3 Top Level definitions
